% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{case_study}
\alias{case_study}
\title{Short version of the \code{case_study_FITC}}
\usage{
case_study
}
\description{
shorter version of the \code{case_study_FITC}. Used as an example in shiny app, 
when the user will not load his own count data.
}
\keyword{datasets}
