% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{case_study_all}
\alias{case_study_all}
\title{Case study with two fluorescent dyes}
\usage{
case_study_all
}
\description{
example data extracted from Aklides system and merged into one file. Counts in this file will not fit properly, 
due to the fact that we integrated into the file counts with two different fluorescent dyes used.
}
\keyword{datasets}
