% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{case_study_APC}
\alias{case_study_APC}
\title{Case study for APC dye}
\usage{
case_study_APC
}
\description{
example data extracted from Aklides system. Counts with only APC fluorescent dye were merged.
}
\keyword{datasets}
