% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{case_study_FITC}
\alias{case_study_FITC}
\title{Case study for FITC dye}
\usage{
case_study_FITC
}
\description{
example data extracted from Aklides system. Counts with only FITC fluorescent dye were merged.
}
\keyword{datasets}
