% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_fit.R
\name{compare_fit}
\alias{compare_fit}
\title{Compare_fit}
\usage{
compare_fit(count_list, fitlist = fit_counts(count_list, model = "all"))
}
\arguments{
\item{count_list}{Table with count data. Each count should be in separate column, rows should represent
values of that counts.}

\item{fitlist}{Uses fit_count function to calculate value for each unique observation using different
distribution models.}
}
\value{
Data table with distribution values for each unique observation in count.
}
\description{
Function calculates distribution of each unique observation value taking in account their
number of occurances for each count.
}
