% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_counts}
\alias{example_counts}
\title{short version of the \code{foci_counts_FITC}}
\usage{
example_counts
}
\description{
shorter version of the \code{foci_count_FITC}. Used as an example in shiny app, 
when the user will not load his own count data.
}
\keyword{datasets}
