% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{foci_count}
\alias{foci_count}
\title{example data extracted from Aklides system and merged into one file. Counts in this file will not fit properly, 
due to the fact that we integrated into the file counts with two different fluorescent dyes used.}
\usage{
foci_count_merged
}
\description{
Data of six bad replicates exported from \code{foci_count_merged}
}
\keyword{datasets}
