% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{foci_count_APC}
\alias{foci_count_APC}
\title{foci counts with APC dye}
\usage{
foci_count_APC
}
\description{
example data extracted from Aklides system. Counts with only APC fluorescent dye were merged.
}
\keyword{datasets}
