% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{foci_count_merged}
\alias{foci_count_merged}
\title{example data extracted from Aklidus system and merged into one file}
\usage{
foci_count_merged
}
\description{
Data of six bad replicates exported from \code{foci_count_merged}
}
\keyword{datasets}
