% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all_func.R
\docType{data}
\name{load_all_func}
\alias{load_all_func}
\alias{functions}
\title{Loads all important functions}
\format{An object of class \code{character} of length 10.}
\usage{
functions
}
\description{
Loads all important functions
}
\keyword{datasets}
