% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_fit.R
\name{plot_fitcmp}
\alias{plot_fitcmp}
\title{plot_fitcmp}
\usage{
plot_fitcmp(fitcmp)
}
\arguments{
\item{fitcmp}{You need to input data frame that is created by compare_fit function.}
}
\description{
Creates a plot from counts name and distribution values of each unique observation, which allows to compare
all the distributions and find the most suitable.
}
