% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_counts.R
\name{process_counts}
\alias{process_counts}
\title{Process counts}
\usage{
process_counts(x)
}
\arguments{
\item{x}{\code{data.frame} or \code{matrix}.}
}
\value{
A \code{list} of counts.
}
\description{
Counts data in table-like format ommiting NAs.
}
\examples{
data(example_counts)
process_counts(example_counts)
}
