% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sim_dat}
\alias{sim_dat}
\title{Data created from simulation of NB Poiss}
\usage{
sim_dat
}
\description{
Data was created using the \code{dplyr} and \code{pbapply} package
}
\seealso{
\code{\link[dplyr]{dplyr}}

\code{\link[pbapply]{pbapply}}
}
\keyword{datasets}
