% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{six_replicates_bad}
\alias{six_replicates_bad}
\title{Six bad replicates}
\usage{
six_replicates_bad
}
\description{
Data of six bad replicates exported from \code{example_counts}
}
\keyword{datasets}
