% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_counts}
\alias{validate_counts}
\title{Validate data}
\usage{
validate_counts(x)
}
\arguments{
\item{x}{\code{data.frame} or \code{matrix}.}
}
\value{
An input object.
}
\description{
Validates count data.
}
\details{
Errors if \code{x} has negative values or non-numeric 
values, otherwise \code{TRUE}.
}
\examples{
data(case_study)
process_counts(case_study)
}
