% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrs.R
\name{zinb}
\alias{zinb}
\alias{ZINB}
\alias{rZINB}
\alias{dZINB}
\title{Zero-inflated negative binomial distrbution}
\usage{
rZINB(n, size, mu, r)

dZINB(x, size, mu, r)
}
\arguments{
\item{n}{number of random values to return.}

\item{size}{target for number of successful trials, or dispersion parameter (the shape 
parameter of the gamma mixing distribution). Must be strictly positive, need not be integer..}

\item{mu}{mean.}

\item{r}{probability of excess zeros.}

\item{x}{vector of (non-negative integer) quantiles.}
}
\description{
Density and random generation for the zero-inflated negative binomial 
distribution.
}
\examples{
rZINB(15, 1.9, 0.9, 0.8)
}
\seealso{
Negative binomial distribution: \code{\link[stats]{NegBinomial}}.
}
