% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrs.R
\name{zip}
\alias{zip}
\alias{dZIP}
\alias{rZIP}
\title{Zero-inflated Poisson distrbution}
\usage{
dZIP(x, lambda, p)

rZIP(n, lambda, p)
}
\arguments{
\item{x}{vector of (non-negative integer) quantiles.}

\item{lambda}{vector of (non-negative) means.}

\item{p}{probability of zeros.}

\item{n}{number of random values to return.}
}
\description{
Density and random generation for the zero inflated Poisson distribution.
}
\examples{
df <- data.frame(poisson = rpois(25, 0.3), binomial = rbinom(25, 1, 0.8))
compare_fit(df, fitlist = fit_counts(df, model = "all"))
}
\seealso{
Poisson distribution:
}
