% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countfitteR.R
\docType{package}
\name{countfitteR_gui}
\alias{countfitteR_gui}
\alias{countfitteR}
\alias{countfitteR_gui-package}
\title{countfitteR Graphical User Interface}
\description{
Launches graphical user interface that analyses given count data and
chooses the best performing distribution model
}
\section{Warning }{
 Any ad-blocking software may cause malfunctions.
}

\examples{
countfitteR_gui()
}
\author{
Jaroslaw Chilimoniuk, Stefan Roediger, Michal Burdukiewcz
}
\keyword{Poisson}
\keyword{count}
\keyword{zero-inflated}
