% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_counts}
\alias{example_counts}
\title{An example dataset}
\usage{
example_counts
}
\description{
Data was created using the healthy list data (not provided in package).
}
\keyword{datasets}
