% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{foci_count_APC}
\alias{foci_count_APC}
\title{example data extracted from Aklidus system. Counts with only APC fluorescent dye were merged.}
\usage{
foci_count_APC
}
\description{
Data of six bad replicates exported from \code{foci_count_APC}.
}
\keyword{datasets}
