% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{foci_count_FITC}
\alias{foci_count_FITC}
\title{example data extracted from Aklides system. Counts with only FITC fluorescent dye were merged.}
\usage{
foci_count_FITC
}
\description{
Data of six bad replicates exported from \code{foci_count_FITC}
}
\keyword{datasets}
