% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{foci_count_FITC_short}
\alias{foci_count_FITC_short}
\title{short version \code{foci_counts_FITC}}
\usage{
foci_count_FITC_short
}
\description{
shorter version of the \code{foci_count_FITC}. Used as an example in shiny app, 
when the user will not load his own count data.
}
\keyword{datasets}
