% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_counts.R
\name{process_counts}
\alias{process_counts}
\title{Process counts}
\usage{
process_counts(x)
}
\arguments{
\item{x}{\code{data.frame} or \code{matrix}.}
}
\value{
A \code{list} of counts.
}
\description{
Converts data in a table-like formats into lists of counts.
}
\details{
\code{case_study} does not consider \code{NA}s and \code{NaN}s effectively 
ommiting them (as the \code{\link[base]{is.na}} function).
}
\examples{
data(case_study)
process_counts(case_study)
}
