% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitlist.R
\name{summary_fitlist}
\alias{summary_fitlist}
\title{summary_fitlist}
\usage{
summary_fitlist(fitlist)
}
\arguments{
\item{fitlist}{List is created by fitlist = fit_counts(count_list, model = "all"). It takes all the counts and for each count
it creates distribution model}
}
\value{
Data frame with summarised results of all distribution models
}
\description{
Function that creates summarized table with parameters of distribution model for each count. Those parameters
are: lambda, lower and upper confidence intervals, BIC, theta and r.
}
